Reversion to ORIGINAL Friends Code
Reverting to the original friends code due to testing revealing my code not working, the code was only written and never tested originally, and now the code actually works as intended. For this reason, the code will need to be re-uploaded to compare to the original, which means it need to be put back into the original code again to compare. Hence this change.